Revert "fix merge conflict?"
This reverts commit 2d648642dff068ae05344ff63860d278f06058ee, reversing
changes made to a02f6627ba1f37a04fabd2b9a1cf06fed48546a3.

+++ source/MenuPanel.h
@@ -55,7 +55,7 @@ private:
 	const Interface *mainMenuUi;
 
 	std::vector<std::string> credits;
-	long long int scroll = 0;
+	unsigned scroll = 0U;
 	bool scrollingPaused = false;
 };
